Move l2 block source to service, replace database dep with param port
